% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_arc-layer.R
\name{add_arc_layer}
\alias{add_arc_layer}
\title{Add arc layer to the deckgl widget}
\usage{
add_arc_layer(deckgl, id = "arc-layer", data = NULL, properties = list(),
  ...)
}
\description{
Add arc layer to the deckgl widget
}
