% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{add_basemap}
\alias{add_basemap}
\title{Add a basemap to the deckgl widget}
\usage{
add_basemap(deckgl, style = use_carto_style(), ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{style}{The style definition of the map conforming to the Mapbox Style Specification.}

\item{...}{not used}
}
\description{
Add a basemap to the deckgl widget
}
