% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_grid-cell-layer.R
\name{add_grid_cell_layer}
\alias{add_grid_cell_layer}
\title{Add a grid cell layer to the deckgl widget}
\usage{
add_grid_cell_layer(deckgl, id = "grid-cell-layer", data = NULL,
  properties = list(), ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{data}{url to fetch data from or data object}

\item{properties}{named list of properties with names corresponding to the properties defined
in the \href{https://deck.gl/#/documentation/deckgl-api-reference}{deckgl-api-reference}
for the given layer class, additionally there is a \code{getTooltip} property (callback)
showing a tooltip when the mouse enters an object,
e. g. \code{getTooltip = JS("object => object.name")}}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
The \code{GridCellLayer} can render a grid-based heatmap.
It is a variation of the \code{ColumnLayer}. It takes the constant width / height of all cells and top-left coordinate of each cell.
The grid cells can be given a height using the \code{getElevation} accessor.
}
\examples{
data("bart_stations")

properties <- list(
  extruded = TRUE,
  cellSize = 50,
  getPosition = ~lng + lat,
  getFillColor = JS("d => [48, 128, Math.min(200, d.spaces * 10)]"),
  getElevation = ~spaces
)

map <- deckgl(zoom = 11, pitch = 45) \%>\%
  add_grid_cell_layer(data = sf_bike_parking, properties = properties) \%>\%
  add_mapbox_basemap()

if (interactive()) map
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/layers/grid-cell-layer}
}
