% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-editor.R
\name{add_json_editor}
\alias{add_json_editor}
\title{Add a JSON-Editor to the deckgl widget}
\usage{
add_json_editor(deckgl, ...)
}
\arguments{
\item{deckgl}{A deckgl widget object.}

\item{...}{Optional args that are passed to the \code{JSONEditor}.}
}
\description{
Adds a JSON-Editor below the map to interact with the layers of your \code{deck} instance.
}
