% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{add_legend}
\alias{add_legend}
\title{Add a legend to the map}
\usage{
add_legend(deckgl, colors, labels, title = NULL, style = NULL, ...)
}
\arguments{
\item{deckgl}{A deckgl widget object.}

\item{colors}{The colors of the legend items.}

\item{labels}{The labels corresponding to the colors of the legend items.}

\item{title}{The title of the legend.}

\item{style}{A \code{cssText} string that will modefiy the default style of the element.}

\item{...}{not used}
}
\description{
Add a legend to the map
}
