% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_pal.R
\name{add_legend_pal}
\alias{add_legend_pal}
\title{Add a legend to the deckgl widget using a palette func}
\usage{
add_legend_pal(deckgl, pal, ...)
}
\arguments{
\item{deckgl}{A deckgl widget object.}

\item{pal}{A palette function that is used to create the
legend elements (colors and labels) automatically.}

\item{...}{Parameters that are passed to \code{\link{add_legend}}.}
}
\description{
Add a legend to the deckgl widget using a palette func
}
\seealso{
\code{\link[scales]{col_numeric}} et cetera for how to create a palette function.
}
