% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl_basemap.R
\name{add_mapbox_basemap}
\alias{add_mapbox_basemap}
\title{Add a base map from mapbox to the deckgl widget}
\usage{
add_mapbox_basemap(deckgl, token = Sys.getenv("MAPBOX_API_TOKEN"),
  style = "mapbox://styles/mapbox/light-v9")
}
\arguments{
\item{deckgl}{deckgl widget}

\item{token}{mapbox api access token}

\item{style}{map style}
}
\value{
deckgl widget
}
\description{
Add a base map from mapbox to the deckgl widget
}
