% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_point-cloud-layer.R
\name{add_point_cloud_layer}
\alias{add_point_cloud_layer}
\title{Add a point cloud layer to the deckgl widget}
\usage{
add_point_cloud_layer(deckgl, id = "point-cloud-layer", data = NULL,
  properties = list(), ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{data}{url to fetch data from or data object}

\item{properties}{named list of properties with names corresponding to the properties defined
in the \href{https://deck.gl/#/documentation/deckgl-api-reference}{deckgl-api-reference}
for the given layer class, additionally there is a \code{getTooltip} property (callback)
showing a tooltip when the mouse enters an object,
e. g. \code{getTooltip = JS("object => object.name")}}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
The \code{PointCloudLayer} takes in points with 3d positions, normals and colors and renders them as spheres with a certain radius.
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/layers/point-cloud-layer}
}
