% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-tiles.R
\name{add_raster_tiles}
\alias{add_raster_tiles}
\title{Add raster tiles layer to the deckgl widget}
\usage{
add_raster_tiles(deckgl, id = "raster-tiles", ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
Add raster tiles layer to the deckgl widget
}
