% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-source.R
\name{add_source_as_dep}
\alias{add_source_as_dep}
\title{Add source as JavaScript dep}
\usage{
add_source_as_dep(deckgl, id, data)
}
\arguments{
\item{deckgl}{A deckgl widget object.}

\item{id}{The unique id of the source.}

\item{data}{The url to fetch data from or a data object.}
}
\description{
Add source as JavaScript dep
}
