% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bart_segments}
\alias{bart_segments}
\title{bart segments}
\format{tibble with 45 rows and 8 variables:
\describe{
  \item{inbound}{number of inbound trips}
  \item{outbound}{number of outbound trips}
  \item{from_name}{source station name}
  \item{from_lng}{source station longitude}
  \item{from_lat}{source station latitude}
  \item{to_name}{target station name}
  \item{to_lng}{target station longitude}
  \item{to_lat}{target station latitude}
}}
\source{
\url{"https://raw.githubusercontent.com/uber-common/deck.gl-data/master/website/bart-segments.json"}
}
\usage{
bart_segments
}
\description{
bart segments
}
\keyword{datasets}
