% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bart_stations}
\alias{bart_stations}
\title{bart stations}
\format{tibble with 44 rows and 7 variables:
\describe{
  \item{name}{station name}
  \item{code}{two-letter station code}
  \item{address}{address}
  \item{entries}{number of entries}
  \item{exits}{number of exits}
  \item{lng}{longitude}
  \item{lat}{latitude}
}}
\source{
\url{https://raw.githubusercontent.com/uber-common/deck.gl-data/master/website/bart-stations.json}
}
\usage{
bart_stations
}
\description{
bart stations
}
\keyword{datasets}
