% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{deckgl_proxy}
\alias{deckgl_proxy}
\title{Create a deckgl proxy object}
\usage{
deckgl_proxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
deck to modify}

\item{session}{the \code{Shiny} session object to which the deckgl widget belongs;
usually the default value will suffice.}
}
\description{
Creates a deckgl-like object that can be used to update a deckgl object that has already been rendered.
}
