% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples_does-it-work.R
\name{does_it_work}
\alias{does_it_work}
\title{Check if everything works fine}
\usage{
does_it_work(token = NULL)
}
\arguments{
\item{token}{mapbox API access token}
}
\description{
Check if everything works fine
}
