% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{encode_icon_atlas}
\alias{encode_icon_atlas}
\title{Encode atlas image to base64}
\usage{
encode_icon_atlas(filename = NULL)
}
\arguments{
\item{filename}{The filename of the atlas image.}
}
\value{
base64 encoded atlas image
}
\description{
Encode atlas image to base64
}
