% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dependency.R
\name{get_data}
\alias{get_data}
\title{Get data}
\usage{
get_data(var_name = "thanksForAllTheFish")
}
\arguments{
\item{var_name}{JavaScript variable name}
}
\description{
EXPERIMENTAL, usually used in conjunction with \link{add_data}
}
