% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties2.R
\name{get_last_element}
\alias{get_last_element}
\title{Create a data accessor retrieving the last element of an array}
\usage{
get_last_element(property_name)
}
\arguments{
\item{property_name}{property name of data object}
}
\value{
JavaScript code evaluated on the client-side
}
\description{
Create a data accessor retrieving the last element of an array
}
