% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{get_property}
\alias{get_property}
\title{Create a data accessor}
\usage{
get_property(property_name)
}
\arguments{
\item{property_name}{property name of data object}
}
\value{
JavaScript code evaluated on the client-side
}
\description{
Creates a JS method to retrieve a given property of each object.
}
