% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{get_value}
\alias{get_value}
\title{Create a data accessor}
\usage{
get_value(property)
}
\arguments{
\item{property}{property of data object}
}
\value{
JavaScript code evaluated on the client-side
}
\description{
Method called to retrieve the value of the given property of each object.
}
