% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-state.R
\name{set_view_state}
\alias{set_view_state}
\title{Set the view state of the map}
\usage{
set_view_state(
  deckgl,
  latitude = 37.8,
  longitude = -122.45,
  zoom = 12,
  pitch = 0,
  bearing = 0
)
}
\arguments{
\item{deckgl}{A deckgl widget object.}

\item{latitude}{The latitude of the view state.}

\item{longitude}{The longitude of the view state.}

\item{zoom}{The zoom level of the view state.}

\item{pitch}{The pitch of the view state.}

\item{bearing}{The bearing of the view state.}
}
\description{
Set the view state of the map
}
