% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{update_deckgl}
\alias{update_deckgl}
\title{Send commands to a deckgl instance in a \code{Shiny} app}
\usage{
update_deckgl(proxy, ...)
}
\arguments{
\item{proxy}{deckgl proxy object}

\item{...}{unused}
}
\description{
Send commands to a deckgl instance in a \code{Shiny} app
}
\seealso{
\link{deckgl_proxy}
}
