% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{use_carto_style}
\alias{use_carto_style}
\title{Use a Carto style}
\usage{
use_carto_style(theme = "dark-matter")
}
\arguments{
\item{theme}{The theme of the style, \code{dark-matter}, \code{positron} or \code{voyager}.}
}
\description{
Use a Carto style
}
