% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_contour-layer.R
\name{use_contour_definition}
\alias{use_contour_definition}
\title{Use contour definition}
\usage{
use_contour_definition(threshold = 1, color = c(255, 255, 255),
  strokeWidth = 1)
}
\arguments{
\item{threshold}{threshold value to be used in contour generation}

\item{color}{RGB color array to be used to render contour lines}

\item{strokeWidth}{width of the contour lines in pixels}
}
\description{
Use contour definition
}
