% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{use_icon_definition}
\alias{use_icon_definition}
\title{Create an icon definition on an atlas image}
\usage{
use_icon_definition(x = 0, y = 0, width = 128, height = 128,
  anchor_x = (width/2), anchor_y = 128, mask = TRUE)
}
\arguments{
\item{x}{The x position of the icon on the atlas image.}

\item{y}{The y position of the icon on the atlas image.}

\item{width}{The width of the icon on the atlas image.}

\item{height}{The height of the icon on the atlas image.}

\item{anchor_x}{The horizontal position of the icon anchor.}

\item{anchor_y}{the vertical position of the icon anchor.}

\item{mask}{whether icon is treated as a transparency mask,
if \code{TRUE}, user defined color is applied,
if \code{FALSE}, pixel color from the image is applied}
}
\description{
Create an icon definition on an atlas image
}
