% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{use_tooltip}
\alias{use_tooltip}
\title{Create a tooltip property}
\usage{
use_tooltip(html, style, ...)
}
\arguments{
\item{html}{The \code{innerHTML} of the element.}

\item{style}{A \code{cssText} string that will modefiy the default style of the element.}

\item{...}{not used}
}
\description{
Create a tooltip property
}
