% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wellknown.R
\name{wellknown}
\alias{wellknown}
\alias{wellknown.sfc_MULTIPOLYGON}
\alias{wellknown.sfc_MULTILINESTRING}
\alias{wellknown.sfc_MULTIPOINT}
\alias{wellknown.default}
\title{Parse geometry column of \code{sf} object to wellknown text}
\usage{
wellknown(data)

\method{wellknown}{sfc_MULTIPOLYGON}(data)

\method{wellknown}{sfc_MULTILINESTRING}(data)

\method{wellknown}{sfc_MULTIPOINT}(data)

\method{wellknown}{default}(data)
}
\arguments{
\item{data}{object of class \code{sf}}
}
\description{
Parse geometry column of \code{sf} object to wellknown text
}
\seealso{
\code{\link{get_wellknown}}
}
