% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_arc-layer.R
\name{add_arc_layer}
\alias{add_arc_layer}
\title{Add an arc layer to the deckgl widget}
\usage{
add_arc_layer(deckgl, id = "arc-layer", data = NULL, properties = list(),
  ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{data}{url to fetch data from or data object}

\item{properties}{named list of properties with names corresponding to the properties defined
in the \href{https://deck.gl/#/documentation/deckgl-api-reference}{deckgl-api-reference}
for the given layer class}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
Add an arc layer to the deckgl widget
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/layers/arc-layer}
}
