% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dependency.R
\name{add_data}
\alias{add_data}
\title{Add JavaScript data file}
\usage{
add_data(deckgl, data, var_name = "thanksForAllTheFish")
}
\arguments{
\item{deckgl}{deckgl widget}

\item{data}{data object}

\item{var_name}{JavaScript variable name used to make the data available}
}
\description{
EXPERIMENTAL
}
