% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_icon-layer.R
\name{add_icon_layer}
\alias{add_icon_layer}
\title{Add an icon layer to the deckgl widget}
\usage{
add_icon_layer(deckgl, id = "icon-layer", data = NULL,
  properties = list(), ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{data}{url to fetch data from or data object}

\item{properties}{named list of properties with names corresponding to the properties defined
in the \href{https://deck.gl/#/documentation/deckgl-api-reference}{deckgl-api-reference}
for the given layer class, additionally there is a \code{getTooltip} property (callback)
showing a tooltip when the mouse enters an object,
e. g. \code{getTooltip = JS("object => object.name")}}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
The Icon Layer renders raster icons at given coordinates.
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/layers/icon-layer}
}
