% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_icon-layer.R
\name{add_icon_layer}
\alias{add_icon_layer}
\title{Add an icon layer to the deckgl widget}
\usage{
add_icon_layer(deckgl, id = "icon-layer", data = NULL,
  properties = default_icon_properties(), ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{data}{url to fetch data from or data object}

\item{properties}{named list of properties with names corresponding to the properties defined
in the \href{https://deck.gl/#/documentation/deckgl-api-reference}{deckgl-api-reference}
for the given layer class, additionally there is a \code{getTooltip} property (callback)
showing a tooltip when the mouse enters an object,
e. g. \code{getTooltip = JS("object => object.name")}}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
The \code{IconLayer} renders raster icons at given coordinates.
}
\examples{
## @knitr icon-layer
sample_data <- paste0(
  "https://raw.githubusercontent.com/",
  "uber-common/deck.gl-data/",
  "master/website/bart-stations.json"
)

properties <- list(
  pickable = TRUE,
  iconAtlas = encode_icon_atlas(),
  iconMapping = list(marker = icon_definition()),
  sizeScale = 10,
  getPosition = get_property("coordinates"),
  getIcon = JS("d => 'marker'"),
  getSize = 5,
  getColor = JS("d => [Math.sqrt(d.exits), 140, 0]"),
  getTooltip = JS("object => `${object.name}<br/>${object.address}`")
)

deck <- deckgl(zoom = 10, pitch = 45) \%>\%
  add_icon_layer(data = sample_data, properties = properties) \%>\%
  add_mapbox_basemap()

if (interactive()) deck
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/layers/icon-layer}
}
