% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_line-layer.R
\name{add_line_layer}
\alias{add_line_layer}
\title{Add a line layer to the deckgl widget}
\usage{
add_line_layer(deckgl, id = "line-layer", data = NULL,
  properties = list(), ...)
}
\arguments{
\item{deckgl}{deckgl widget}

\item{id}{id of the layer}

\item{data}{url to fetch data from or data object}

\item{properties}{named list of properties with names corresponding to the properties defined
in the \href{https://deck.gl/#/documentation/deckgl-api-reference}{deckgl-api-reference}
for the given layer class, additionally there is a \code{getTooltip} property (callback)
showing a tooltip when the mouse enters an object,
e. g. \code{getTooltip = JS("object => object.name")}}

\item{...}{more properties (will be added to the properties object), useful if you
want to use a properties object for more than one layer}
}
\description{
The \code{LineLayer} renders flat lines joining pairs of source and target points,
specified as latitude/longitude coordinates.
}
\examples{
## @knitr line-layer
data("bart_segments")

properties <- list(
  pickable = TRUE,
  getWidth = 12,
  getSourcePosition = ~from_lng + from_lat,
  getTargetPosition = ~to_lng + to_lat,
  getColor = JS("d => [Math.sqrt(d.inbound + d.outbound), 140, 0]"),
  getTooltip = JS("object => `${object.from_name} to ${object.to_name}`")
)

deck <- deckgl(zoom = 10, pitch = 20) \%>\%
  add_line_layer(data = bart_segments, properties = properties) \%>\%
  add_mapbox_basemap()

if (interactive()) deck
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/layers/line-layer}
}
