% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapboxgl_basemap.R
\name{add_mapbox_basemap}
\alias{add_mapbox_basemap}
\title{Add a basemap from mapbox to the deckgl widget}
\usage{
add_mapbox_basemap(deckgl, style = "mapbox://styles/mapbox/light-v9",
  token = Sys.getenv("MAPBOX_API_TOKEN"))
}
\arguments{
\item{deckgl}{deckgl widget}

\item{style}{map style}

\item{token}{mapbox API access token}
}
\value{
deckgl widget
}
\description{
Add a basemap from mapbox to the deckgl widget
}
