% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bart_segments}
\alias{bart_segments}
\title{bart segments}
\format{tibble with 45 rows and 8 variables:
\describe{
  \item{inbound}{number of inbound trips}
  \item{outbound}{number of outbound trips}
  \item{from_name}{name of source station }
  \item{from_lng}{longitude of source station}
  \item{from_lat}{latitude of source station}
  \item{to_name}{name of target station}
  \item{to_lng}{longitude of target station}
  \item{to_lat}{latitude of target station}
}}
\source{
\url{https://raw.githubusercontent.com/uber-common/deck.gl-data/master/website/bart-segments.json}
}
\usage{
bart_segments
}
\description{
bart segments
}
\keyword{datasets}
