% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deckgl.R
\name{deckgl}
\alias{deckgl}
\title{Create a deckgl widget}
\usage{
deckgl(latitude = 37.8, longitude = -122.45, zoom = 12, pitch = 0,
  bearing = 0, initial_view_state = NULL, views = NULL,
  width = NULL, height = NULL, element_id = NULL, ...)
}
\arguments{
\item{latitude}{The latitude of the initial view state.}

\item{longitude}{The longitude of the initial view state.}

\item{zoom}{The zoom level of the initial view state.}

\item{pitch}{The pitch of the initial view state.}

\item{bearing}{The bearing of the initial view state.}

\item{initial_view_state}{The initial view state.
If set, other view state arguments (\code{longitude}, \code{latidude} et cetera) are ignored.}

\item{views}{A single \code{View}, or an array of \code{View} instances.
If not supplied, a single \code{MapView} will be created.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{element_id}{The explicit id of the widget (usually not needed).}

\item{...}{Optional properties that are passed to the \code{deck} instance.}
}
\value{
deckgl widget
}
\description{
Create a deckgl widget
}
\seealso{
\url{https://deck.gl/#/documentation/deckgl-api-reference/deck} for optional properties
  that can be passed to the \code{deck} instance.
}
