% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{get_color_to_rgb_array}
\alias{get_color_to_rgb_array}
\title{Create a getColor data accessor}
\usage{
get_color_to_rgb_array(color_property)
}
\arguments{
\item{color_property}{property name of data object containing the HEX color}
}
\value{
JavaScript code evaluated on the client-side
}
\description{
Creates a JS method to retrieve the color of each object.
The method parses the HEX color property of the data object to an rgb color array.
}
