% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{get_position}
\alias{get_position}
\title{Create a getPosition property}
\usage{
get_position(latitude = NULL, longitude = NULL, coordinates = NULL)
}
\arguments{
\item{latitude}{latitude key in data object}

\item{longitude}{longitude key in data object}

\item{coordinates}{coordinates key in data object
(in this case \code{latitude} and \code{longitude} parameters are ignored)}
}
\value{
getPosition property
}
\description{
Create a getPosition property
}
