% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{JS}
\alias{onRender}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmlwidgets}{\code{\link[htmlwidgets]{JS}}, \code{\link[htmlwidgets]{onRender}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

