% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_contour-layer.R
\name{use_contour_definition}
\alias{use_contour_definition}
\title{Create a contour definition}
\usage{
use_contour_definition(threshold = 1, color = c(255, 255, 255),
  stroke_width = 1)
}
\arguments{
\item{threshold}{The threshold value used in contour generation.}

\item{color}{The RGB color array used to render contour lines.}

\item{stroke_width}{The width of the contour lines in pixels.}
}
\description{
Create a contour definition
}
