% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{use_icon_definition}
\alias{use_icon_definition}
\title{Icon definition on an atlas image}
\usage{
use_icon_definition(x = 0, y = 0, width = 128, height = 128,
  anchorX = (width/2), anchorY = 128, mask = TRUE)
}
\arguments{
\item{x}{x position of icon on the atlas image}

\item{y}{y position of icon on the atlas image}

\item{width}{width of icon on the atlas image}

\item{height}{height of icon on the atlas image}

\item{anchorX}{horizontal position of icon anchor}

\item{anchorY}{vertical position of icon anchor}

\item{mask}{whether icon is treated as a transparency mask,
if \code{TRUE}, user defined color is applied,
if \code{FALSE}, pixel color from the image is applied}
}
\description{
Icon definition on an atlas image
}
